Edit: 메뉴별 판매 리포트 썸네일 디자인 수정 및 구조 수정
