+++ src/components/ui/MenuElement.tsx
@@ -15,17 +15,19 @@ const MenuElement = ({ ...props }: MenuElementProps) => {
         (props.className ?? "")
       }
     >
-      <div className="flex w-6 items-center justify-center">{props.number}</div>
+      <div className="flex items-center justify-center w-6 font-medium">
+        {props.number}
+      </div>
       <div className={"flex w-6 h-6 items-center justify-center"}>
         <img
           src={props.image}
           alt="메뉴 이미지"
-          className="rounded-full object-cover border border-neutral-300 "
+          className="object-cover border rounded-full border-neutral-300 "
         />
       </div>
 
       <div className="flex-grow">{props.title}</div>
-      <div>{props.paragraph}</div>
+      <div className="font-medium">{props.paragraph}</div>
     </div>
   );
 };
